Move Scheduler to Core
