+++ PSXSharp/Core/Interpreter/CPU_Interpreter.cs
@@ -1,9 +1,7 @@
 ﻿using PSXSharp.Core.Common;
-using PSXSharp.Core.x64_Recompiler;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Text;
 using static PSXSharp.Core.CPU;
