+++ PSXSharp/Peripherals/IO/JOY.cs
@@ -1,5 +1,4 @@
 ﻿using PSXSharp.Core;
-using PSXSharp.Core.x64_Recompiler;
 using System;
 using System.IO;
