+++ PSXSharp/Peripherals/Timers/Timer1.cs
@@ -1,5 +1,4 @@
 ﻿using PSXSharp.Core;
-using PSXSharp.Core.x64_Recompiler;
 using System;
 using static PSXSharp.CD_ROM;
