+++ PSXSharp/Peripherals/Timers/Timer2.cs
@@ -1,5 +1,4 @@
 ﻿using PSXSharp.Core;
-using PSXSharp.Core.x64_Recompiler;
 using System;
 
 namespace PSXSharp.Peripherals.Timers {
